server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
numpy==1.26.3
open-clip-torch==2.24.0
torch==2.1.2
